requirements.txt
South==0.7.3
django-crispy-forms==1.1.1
django-debug-toolbar==0.9.4
django-grappelli==2.3.7
django-pagination==1.0.7
django-coverage==1.2.2
coverage==3.5.1
geopy==0.94.1
ipython==0.11
mock==0.7.2
psycopg2==2.4.4
sentry==3.3.1
gunicorn==0.13.4
# django development version
git+git://github.com/django/django.git

